refactor: move storage & redis into its own crates
