+++ crates/storage/src/lib.rs
@@ -1,0 +1,23 @@
+pub mod redis;
+pub mod rocksdb;
+
+
+#[auto_impl(&, Box, Arc)]
+pub trait Database: Send + Sync + TreeReader + TreeWriter {
+    fn get_hashchain(&self, key: &str) -> Result<Hashchain>;
+    fn set_hashchain(&self, incoming_operation: &Operation, value: &[HashchainEntry])
+        -> Result<()>;
+
+    fn get_commitment(&self, epoch: &u64) -> Result<String>;
+    fn set_commitment(&self, epoch: &u64, commitment: &Digest) -> Result<()>;
+
+    fn get_epoch(&self) -> Result<u64>;
+    fn set_epoch(&self, epoch: &u64) -> Result<()>;
+
+    #[cfg(test)]
+    fn flush_database(&self) -> Result<()>;
+}
+
+fn convert_to_connection_error(e: redis::RedisError) -> PrismError {
+    PrismError::Database(DatabaseError::ConnectionError(e.to_string()))
+}
